created a separate route/controller file for categories API
