+++ controllers/inventory-controller.js
@@ -61,19 +61,4 @@ const getItemDetailsById = async (req, res) => {
   }
 };
 
-// get all the inventory categories
-const getInventoryCategories = async (req, res) => {
-  try {
-    const response = await knex("inventories")
-      .pluck("category")
-      .distinct("category");
-
-    res.status(200).json(response);
-  } catch (error) {
-    res.status(500).json({
-      message: `Unable to retrieve inventory categories`,
-    });
-  }
-}
-
-export { getInventoryByWarehouseId, getItemDetailsById, getAllInventory, getInventoryCategories };
+export { getInventoryByWarehouseId, getItemDetailsById, getAllInventory };
