+++ routes/inventory.js
@@ -6,13 +6,8 @@ const router = express.Router();
 // usage: GET /api/inventories
 router.route("/").get(inventoryController.getAllInventory);
 
-// get all categories for inventory
-// usage: GET /api/inventories/categories
-router.route("/categories").get(inventoryController.getInventoryCategories);
-
 // get item details based on ID
 // usage: GET /api/inventories/:id
 router.route("/:id").get(inventoryController.getItemDetailsById);
-
 
 export default router;
